remove dependencies to fedbiomed.node in fedbiomed.common - for v6.0.0
