+++ fedbiomed/common/cli.py
@@ -8,20 +8,17 @@ This module includes common CLI methods and parser extension
 """
 
 import argparse
-import importlib
 import os
 import sys
 from abc import ABC, abstractmethod
 from typing import Dict, List
 
 from fedbiomed.common.certificate_manager import CertificateManager
-from fedbiomed.common.config import Config, docker_special_case
+from fedbiomed.common.config import Config
 from fedbiomed.common.constants import (
     CONFIG_FOLDER_NAME,
     DB_FOLDER_NAME,
     ComponentType,
-    DEFAULT_NODE_NAME,
-    DEFAULT_RESEARCHER_NAME,
     __version__,
 )
 from fedbiomed.common.exceptions import FedbiomedError
@@ -39,15 +36,6 @@ YLW = "\033[1;33m"  # yellow
 GRN = "\033[1;32m"  # green
 NC = "\033[0m"  # no color
 BOLD = "\033[1m"
-
-
-class UniqueStore(argparse.Action):
-    """Argparse action for avoiding having several time the same optional
-      argument"""
-    def __call__(self, parser, namespace, values, option_string):
-        if getattr(namespace, self.dest, self.default) is not self.default:
-            parser.error(option_string + " appears several times.")
-        setattr(namespace, self.dest, values)
 
 
 class CLIArgumentParser:
@@ -134,132 +122,6 @@ class ComponentDirectoryAction(ABC, argparse.Action):
         logger.setLevel("DEBUG")
 
 
-class ComponentParser(CLIArgumentParser):
-    """Instantiates configuration parser"""
-
-    def initialize(self):
-        """Initializes argument parser for creating configuration file."""
-
-        self._parser = self._subparser.add_parser(
-            "component",
-            help="The helper for generating or updating component configuration files, see `configuration -h`"
-            " for more details",
-        )
-
-        self._parser.set_defaults(func=self.default)
-
-        # Common parser to register common arguments for create and refresh
-        common_parser = argparse.ArgumentParser(add_help=False)
-        common_parser.add_argument(
-            "-p",
-            "--path",
-            action=UniqueStore,
-            metavar="COMPONENT_PATH",
-            type=str,
-            nargs="?",
-            required=False,
-            help="Path to specificy where Fed-BioMed component will be intialized.",
-        )
-
-        common_parser.add_argument(
-            "-c",
-            "--component",
-            metavar="COMPONENT_TYPE[ NODE|RESEARCHER ]",
-            type=str,
-            nargs="?",
-            required=True,
-            help="Component type NODE or RESEARCHER",
-        )
-
-        # Create sub parser under `configuration` command
-        component_sub_parsers = self._parser.add_subparsers()
-
-        create = component_sub_parsers.add_parser(
-            "create",
-            parents=[common_parser],
-            help="Creates component folder for the specified component if it does not exist. "
-            "If the component folder exists, leave it unchanged",
-        )
-
-        create.add_argument(
-            "-eo",
-            "--exist-ok",
-            action="store_true",
-            help="Creates configuration only if there isn't an existing one",
-        )
-
-        create.set_defaults(func=self.create)
-
-    def _get_component_instance(self, path: str, component: str):
-        """Gets component"""
-        if component.lower() == "node":
-            config_node = importlib.import_module("fedbiomed.node.config")
-            _component = config_node.node_component
-        elif component.lower() == "researcher":
-            os.environ["FBM_RESEARCHER_COMPONENT_ROOT"] = path
-            config_researcher = importlib.import_module(
-                "fedbiomed.researcher.config"
-            )
-            _component = config_researcher.researcher_component
-        else:
-            print(f"Undefined component type {component}")
-            sys.exit(101)
-        
-        return _component
-
-    def create(self, args):
-        """CLI Handler for creating configuration file and assets for given component
-        """
-        if not args.path:
-            if args.component.lower() == "researcher":
-                component_path = os.path.join(os.getcwd(), DEFAULT_RESEARCHER_NAME)
-            else:
-                component_path = os.path.join(os.getcwd(), DEFAULT_NODE_NAME)
-        else:
-            component_path = args.path
-
-        # Researcher specific case ----------------------------------------------------
-        # This is a special case since researcher import
-        if args.component is None:
-            CommonCLI.error("Error: bad command line syntax")
-        if args.component.lower() == "researcher":
-            if DEFAULT_RESEARCHER_NAME in component_path and \
-                os.path.isdir(component_path) and \
-                not docker_special_case(component_path):
-                if not args.exist_ok:
-                    CommonCLI.error(
-                        f"Default component is already existing. In the directory {component_path} "
-                        "please remove existing one to re-initiate"
-                    )
-                    sys.exit(1)
-                else:
-                    CommonCLI.success(
-                        "Component is already existing. Using existing component."
-                    )
-                    sys.exit(0)
-            else:
-                self._get_component_instance(component_path, args.component)
-                return
-        else:
-            component = self._get_component_instance(component_path, args.component)
-            # Overwrite force configuration file
-            if component.is_component_existing(component_path):
-                if not args.exist_ok:
-                    CommonCLI.error(
-                        f"Component is already existing in the directory `{component_path}`. To ignore "
-                       "this error please execute component creation using `--exist-ok`"
-                )
-                else:
-                    CommonCLI.success(
-                        "Component is already exsiting. Using existing component."
-                    )
-                    return
-
-            component.initiate(component_path)
-
-        CommonCLI.success(f"Component has been initialized in {component_path}")
-
-
 class CommonCLI:
 
     _arg_parsers_classes: List[type] = []
@@ -276,9 +138,6 @@ class CommonCLI:
         self._certificate_manager: CertificateManager = CertificateManager()
         self._description: str = ""
         self._args = None
-
-        # Initialize configuration parser
-        self.configuration_parser = ComponentParser(self._subparsers)
 
     @property
     def parser(self) -> argparse.ArgumentParser:
@@ -367,7 +226,6 @@ class CommonCLI:
         is not executed.
         """
 
-        self.configuration_parser.initialize()
         self.initialize_magic_dev_environment_parsers()
         self.initialize_version()
