+++ tests/test_common_cli.py
@@ -1,63 +1,11 @@
-import argparse
 import shutil
 import sys
 import tempfile
 import unittest
 from unittest.mock import MagicMock, patch
 
-from fedbiomed.common.cli import CommonCLI, ComponentParser
+from fedbiomed.common.cli import CommonCLI
 from fedbiomed.common.exceptions import FedbiomedError
-
-
-class TestComponentParser(unittest.TestCase):
-
-    def setUp(self):
-
-        self.main_parser = argparse.ArgumentParser()
-        self.parser = self.main_parser.add_subparsers()
-        self.conf_parser = ComponentParser(subparser=self.parser)
-        self.conf_parser.initialize()
-
-        self.tem = tempfile.TemporaryDirectory()
-
-
-
-    def tearDown(self):
-        self.tem.cleanup()
-        pass
-
-    def test_01_component_parser_initialize(self):
-        """Tests argument initialization"""
-        self.assertTrue("component" in self.conf_parser._subparser.choices)
-        self.assertTrue(
-            "create"
-            in self.conf_parser._subparser.choices["component"]
-            ._subparsers._group_actions[0]
-            .choices
-        )
-        self.assertEqual(
-            self.conf_parser._subparser.choices["component"]
-            ._subparsers._group_actions[0]
-            .choices["create"]
-            ._defaults["func"]
-            .__func__.__name__,
-            "create",
-        )
-
-    def test_02_component_parser_create(
-        self,
-    ):
-        args = self.main_parser.parse_args(
-            ["component", "create", "--path", self.tem.name, "--component", "NODE", "-eo"]
-        )
-        self.conf_parser.create(args)
-        self.tem.cleanup()
-
-
-        args = self.main_parser.parse_args(
-            ["component", "create", "--path", self.tem.name, "--component", "RESEARCHER", "-eo"]
-        )
-        self.conf_parser.create(args)
 
 
 class TestCommonCLI(unittest.TestCase):
@@ -106,7 +54,6 @@ class TestCommonCLI(unittest.TestCase):
         self.cli.initialize_optional()
 
         self.assertTrue("certificate-dev-setup" in self.cli._subparsers.choices)
-        self.assertTrue("component" in self.cli._subparsers.choices)
 
     def test_04_common_cli_initialize_magic_dev_environment_parsers(self):
         self.cli.initialize_magic_dev_environment_parsers()
